[FEATURE] Create database tables and adjust existing table queries
It is a TYPO3 convention to write CREATE TABLE for SQL tables. It does
not matter if you adjust the tbale or create a new one. TYPO3 change the
SQL query if the table is already there and only care for the difference.

This commit take care of the same behavoir for the testing instance.